rasterization/requirements.txt
asttokens==2.4.1
backcall==0.2.0
decorator==5.1.1
executing==2.1.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==1.24.4
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
tomli==2.0.2
tqdm==4.66.5
traitlets==5.14.3
trimesh==4.5.0
typing_extensions==4.12.2
wcwidth==0.2.13
